requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.3
django-widget-tweaks==1.5.0
idna==3.10
paypalrestsdk==1.13.3
pillow==11.0.0
pycparser==2.22
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
urllib3==2.2.3
